+++ app/templates/index.js
@@ -1,8 +1,10 @@
-import { storiesOf } from 'storybook-react'
-import { action } from '@storybook/addon-actions'
+import { storiesOf, setAddon } from 'storybook-react'
+import JSXAddon from 'storybook-addon-jsx';
 
 import <%= componentCC %> from './src/js/export'
 
+setAddon(JSXAddon)
+
 storiesOf('<%= componentCC %>', module)
-  .add('1 Column', () => <<%= componentCC %> type={'1-column'} />)
-  .add('2 Columns', () => <<%= componentCC %> type={'2-columns'} />)
+  .addWithJSX('1 Column', () => <<%= componentCC %> type={'1-column'} />)
+  .addWithJSX('2 Columns', () => <<%= componentCC %> type={'2-columns'} />)